Add NULL check at opendir trace->directory

+++ include/dlt/dlt_offline_trace.h
@@ -126,7 +126,7 @@ extern DltReturnValue dlt_offline_trace_write(DltOfflineTrace *trace,unsigned ch
  * Get size of currently used offline trace buffer
  * @return size in bytes
  */
-extern unsigned long dlt_offline_trace_get_total_size(DltOfflineTrace *trace);
+extern int dlt_offline_trace_get_total_size(DltOfflineTrace *trace);
 
 /**
  * Provides info about the offline logs storage directory
